+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/StainDto.java
@@ -1,31 +1,42 @@
 package uk.ac.bbsrc.tgac.miso.dto;
 
 public class StainDto {
-  private String category;
-  private long id;
-  private String name;
 
-  public String getCategory() {
-    return category;
+  private Long id;
+  private String name;
+  private Long categoryId;
+  private String categoryName;
+
+  public Long getId() {
+    return id;
   }
 
-  public long getId() {
-    return id;
+  public void setId(Long id) {
+    this.id = id;
   }
 
   public String getName() {
     return name;
   }
 
-  public void setCategory(String category) {
-    this.category = category;
-  }
-
-  public void setId(long id) {
-    this.id = id;
-  }
-
   public void setName(String name) {
     this.name = name;
   }
+
+  public Long getCategoryId() {
+    return categoryId;
+  }
+
+  public void setCategoryId(Long categoryId) {
+    this.categoryId = categoryId;
+  }
+
+  public String getCategoryName() {
+    return categoryName;
+  }
+
+  public void setCategoryName(String categoryName) {
+    this.categoryName = categoryName;
+  }
+
 }